Added some more redundant message checking, but recent monitor changes have introduced an echo chamber bug.
This commit doesn't actually complete the example. We have scaled it down to just one monitor while we learn how to debug rabbitmq messaging. Something is causing the agents to keep receiving the same message over and over, and we need to learn how to ask the broker server what it is seeing.