The deferred blocking channel now clears out the callback queue when it finishes processing them. Whoops!
What happened was it would just keep sending the message, over and over. I never cleared the queue after iterating them. This is safe to do because it has a lock on it.

Also checked in some helper code that I was trying to use to isolate the problem. I am going to keep it around for a few more commits.